+++ spec/rubocop/cop/style/guard_clause_spec.rb
@@ -9,14 +9,14 @@ RSpec.describe RuboCop::Cop::Style::GuardClause, :config do
       expect_offense(<<~RUBY)
         def func
           if something
-          ^^ Use a guard clause instead of wrapping the code inside a conditional expression.
+          ^^ Use a guard clause (`return unless something`) instead of wrapping the code inside a conditional expression.
             #{body}
           end
         end
 
         def func
           unless something
-          ^^^^^^ Use a guard clause instead of wrapping the code inside a conditional expression.
+          ^^^^^^ Use a guard clause (`return if something`) instead of wrapping the code inside a conditional expression.
             #{body}
           end
         end
@@ -28,7 +28,7 @@ RSpec.describe RuboCop::Cop::Style::GuardClause, :config do
         def func
           test
           if something
-          ^^ Use a guard clause instead of wrapping the code inside a conditional expression.
+          ^^ Use a guard clause (`return unless something`) instead of wrapping the code inside a conditional expression.
             #{body}
           end
         end
@@ -36,7 +36,7 @@ RSpec.describe RuboCop::Cop::Style::GuardClause, :config do
         def func
           test
           unless something
-          ^^^^^^ Use a guard clause instead of wrapping the code inside a conditional expression.
+          ^^^^^^ Use a guard clause (`return if something`) instead of wrapping the code inside a conditional expression.
             #{body}
           end
         end
@@ -136,14 +136,14 @@ RSpec.describe RuboCop::Cop::Style::GuardClause, :config do
       expect_offense(<<~RUBY)
         def func
           if something
-          ^^ Use a guard clause instead of wrapping the code inside a conditional expression.
+          ^^ Use a guard clause (`return unless something`) instead of wrapping the code inside a conditional expression.
             work
           end
         end
 
         def func
           unless something
-          ^^^^^^ Use a guard clause instead of wrapping the code inside a conditional expression.
+          ^^^^^^ Use a guard clause (`return if something`) instead of wrapping the code inside a conditional expression.
             work
           end
         end
@@ -200,7 +200,7 @@ RSpec.describe RuboCop::Cop::Style::GuardClause, :config do
     it "registers an error with #{kw} in the if branch" do
       expect_offense(<<~RUBY)
         if something
-        ^^ Use a guard clause instead of wrapping the code inside a conditional expression.
+        ^^ Use a guard clause (`#{kw} if something`) instead of wrapping the code inside a conditional expression.
           #{kw}
         else
           puts "hello"
@@ -211,7 +211,7 @@ RSpec.describe RuboCop::Cop::Style::GuardClause, :config do
     it "registers an error with #{kw} in the else branch" do
       expect_offense(<<~RUBY)
         if something
-        ^^ Use a guard clause instead of wrapping the code inside a conditional expression.
+        ^^ Use a guard clause (`#{kw} unless something`) instead of wrapping the code inside a conditional expression.
          puts "hello"
         else
           #{kw}
@@ -266,7 +266,7 @@ RSpec.describe RuboCop::Cop::Style::GuardClause, :config do
     it 'registers an error if non-control-flow branch has multiple lines' do
       expect_offense(<<~RUBY)
         if something
-        ^^ Use a guard clause instead of wrapping the code inside a conditional expression.
+        ^^ Use a guard clause (`#{kw} if something`) instead of wrapping the code inside a conditional expression.
           #{kw}
         else
           puts "hello" \\
@@ -287,7 +287,7 @@ RSpec.describe RuboCop::Cop::Style::GuardClause, :config do
         module CopTest
           def test
             if something
-            ^^ Use a guard clause instead of wrapping the code inside a conditional expression.
+            ^^ Use a guard clause (`return unless something`) instead of wrapping the code inside a conditional expression.
               work
             end
           end
@@ -299,8 +299,8 @@ RSpec.describe RuboCop::Cop::Style::GuardClause, :config do
       expect_offense(<<~RUBY)
         module CopTest
           def self.test
-            if something
-            ^^ Use a guard clause instead of wrapping the code inside a conditional expression.
+            if something && something_else
+            ^^ Use a guard clause (`return unless something && something_else`) instead of wrapping the code inside a conditional expression.
               work
             end
           end